move position of rewriter.
